+++ CHANGELOG.md
@@ -79,3 +79,4 @@ Features for initial release:
 
 - Add `nullWhenEmpty` extension on `String`
 - Update and restrict dependencies
+- Add `split` extension on any `Iterable` or `List`